remove as property, update imports to use new name, get rid of separate sidebar files
