+++ airbyte-webapp/src/packages/cloud/views/layout/CloudMainView/CloudBottomItems.tsx
@@ -5,6 +5,7 @@ import { CreditsIcon } from "components/icons/CreditsIcon";
 import { FeatureItem, useFeature } from "hooks/services/Feature";
 import { CloudRoutes } from "packages/cloud/cloudRoutePaths";
 import { useGetCloudWorkspace } from "packages/cloud/services/workspaces/CloudWorkspacesService";
+import { LOW_BALANCE_CREDIT_THRESHOLD } from "packages/cloud/views/credits/CreditsPage/components/LowCreditBalanceHint/LowCreditBalanceHint";
 import { RoutePaths } from "pages/routePaths";
 import { useCurrentWorkspace } from "services/workspaces/WorkspacesService";
 import { MenuContent } from "views/layout/SideBar/components/MenuContent";
@@ -13,7 +14,6 @@ import SettingsIcon from "views/layout/SideBar/components/SettingsIcon";
 
 import { CloudResourcesDropdown } from "./CloudResourcesDropdown";
 import { CloudSupportDropdown } from "./CloudSupportDropdown";
-import { LOW_BALANCE_CREDIT_THRESHOLD } from "../../credits/CreditsPage/components/LowCreditBalanceHint/LowCreditBalanceHint";
 
 export const CloudBottomItems: React.FC = () => {
   const workspace = useCurrentWorkspace();
@@ -25,7 +25,6 @@ export const CloudBottomItems: React.FC = () => {
       <MenuContent>
         <li>
           <NavItem
-            as="navLink"
             to={CloudRoutes.Credits}
             icon={<CreditsIcon />}
             label={<FormattedMessage id="sidebar.credits" />}
@@ -41,7 +40,6 @@ export const CloudBottomItems: React.FC = () => {
         </li>
         <li>
           <NavItem
-            as="navLink"
             label={<FormattedMessage id="sidebar.settings" />}
             icon={<SettingsIcon />}
             to={RoutePaths.Settings}
