+++ airbyte-webapp/src/packages/cloud/views/layout/CloudMainView/CloudMainView.module.scss
@@ -4,6 +4,44 @@
 $sidebar-workspace-font-size: 9px;
 $sidebar-workspace-font-weight: 400;
 
+$banner-height: 30px;
+
+.mainContainer {
+  overflow: hidden;
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  height: 100%;
+  min-height: 680px;
+
+  .content {
+    overflow-y: auto;
+    width: 100%;
+    height: 100%;
+
+    .dataBlock {
+      width: 100%;
+      height: 100%;
+    }
+
+    &.alertBanner {
+      margin-top: $banner-height;
+
+      .dataBlock {
+        height: calc(100% - #{$banner-height});
+      }
+    }
+
+    &.speedyConnectionBanner {
+      margin-top: 50px;
+
+      .dataBlock {
+        height: calc(100% - 50px);
+      }
+    }
+  }
+}
+
 .workspaceButton {
   overflow: hidden;
   cursor: pointer;
