+++ airbyte-webapp/src/views/layout/MainView/MainView.module.scss
@@ -1,0 +1,12 @@
+.mainViewContainer {
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+  min-height: 680px;
+}
+
+.content {
+  overflow-y: auto;
+  width: 100%;
+  height: 100%;
+}
