+++ airbyte-webapp/src/views/layout/MainView/MainView.tsx
@@ -1,37 +1,38 @@
 import React from "react";
-import styled from "styled-components";
 
 import { LoadingPage } from "components";
+import { FlexContainer } from "components/ui/Flex";
 
+import { useConfig } from "config";
 import { ResourceNotFoundErrorBoundary } from "views/common/ResorceNotFoundErrorBoundary";
 import { StartOverErrorView } from "views/common/StartOverErrorView";
 
-import { SideBar } from "./SideBar";
+import styles from "./MainView.module.scss";
+import { AirbyteHomeLink } from "../SideBar/AirbyteHomeLink";
+import { BottomItems } from "../SideBar/BottomItems";
+import { MenuContent } from "../SideBar/components/MenuContent";
+import { MainNavItems } from "../SideBar/MainNavItems";
+import { SideBar } from "../SideBar/SideBar";
 
-const MainContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-  display: flex;
-  flex-direction: row;
-  min-height: 680px;
-`;
+const MainView: React.FC<React.PropsWithChildren<unknown>> = (props) => {
+  const { version } = useConfig();
 
-const Content = styled.div`
-  overflow-y: auto;
-  width: 100%;
-  height: 100%;
-`;
-
-const MainView: React.FC<React.PropsWithChildren<unknown>> = (props) => (
-  <MainContainer>
-    <SideBar />
-    <Content>
-      <ResourceNotFoundErrorBoundary errorComponent={<StartOverErrorView />}>
-        <React.Suspense fallback={<LoadingPage />}>{props.children}</React.Suspense>
-      </ResourceNotFoundErrorBoundary>
-    </Content>
-  </MainContainer>
-);
+  return (
+    <FlexContainer className={styles.mainViewContainer}>
+      <SideBar>
+        <AirbyteHomeLink />
+        <MenuContent>
+          <MainNavItems />
+          <BottomItems version={version} />
+        </MenuContent>
+      </SideBar>
+      <div className={styles.content}>
+        <ResourceNotFoundErrorBoundary errorComponent={<StartOverErrorView />}>
+          <React.Suspense fallback={<LoadingPage />}>{props.children}</React.Suspense>
+        </ResourceNotFoundErrorBoundary>
+      </div>
+    </FlexContainer>
+  );
+};
 
 export default MainView;
