+++ airbyte-webapp/src/views/layout/SideBar/NotificationIndicator.module.scss
@@ -1,0 +1,5 @@
+.notification {
+  position: absolute;
+  top: 11px;
+  right: 23px;
+}
