+++ airbyte-webapp/src/views/layout/SideBar/NotificationIndicator.tsx
@@ -1,18 +1,13 @@
 import React from "react";
-import styled from "styled-components";
 
 import Indicator from "components/Indicator";
 
 import { useGetConnectorsOutOfDate } from "hooks/services/useConnector";
 
-const Notification = styled(Indicator)`
-  position: absolute;
-  top: 11px;
-  right: 23px;
-`;
+import styles from "./NotificationIndicator.module.scss";
 
 export const NotificationIndicator: React.FC = () => {
   const { hasNewVersions } = useGetConnectorsOutOfDate();
 
-  return hasNewVersions ? <Notification /> : null;
+  return hasNewVersions ? <Indicator className={styles.notification} /> : null;
 };
