+++ airbyte-webapp/src/views/layout/SideBar/SideBar.tsx
@@ -1,9 +1,9 @@
-import React, { PropsWithChildren } from "react";
+import { PropsWithChildren } from "react";
 
 import { MenuContent } from "./components/MenuContent";
 import styles from "./GenericSideBar.module.scss";
 
-export const GenericSideBar: React.FC<PropsWithChildren<unknown>> = ({ children }) => {
+export const SideBar: React.FC<PropsWithChildren<unknown>> = ({ children }) => {
   return (
     <nav className={styles.nav}>
       <MenuContent>{children}</MenuContent>
